Rename files and find->read
