+++ examples/index.html
@@ -27,7 +27,7 @@
 
 			// Fetch and render some channels to test.
 			const ul = document.querySelector('#channels')
-			channel.findChannels(100).then(response => {
+			channel.readChannels(100).then(response => {
 				console.log(response)
 				for (const c of response.data) {
 					const li = document.createElement('li')
@@ -35,7 +35,7 @@
 					ul.appendChild(li)
 				}
 				// Testing it worked
-				channel.findChannelBySlug(response.data[0].slug).then(res => console.log(res))
+				channel.readChannel(response.data[0].slug).then(res => console.log(res))
 			})
 
 			/**
@@ -49,9 +49,9 @@
 				console.log('testing')
 				const {data: {user}} = await sdk.auth.signIn({email, password})
 				await sdk.channel.createChannel({name: 'Radio Test', slug: slug})
-				const {data: channel} = await sdk.channel.findChannelBySlug(slug)
+				const {data: channel} = await sdk.channel.readChannel(slug)
 				await sdk.channel.updateChannel(channel.id, {name: 'updated'})
-				const {data: updatedChannel} = await sdk.channel.findChannelBySlug(slug)
+				const {data: updatedChannel} = await sdk.channel.readChannel(slug)
 				console.log('same channel', channel.id === updatedChannel.id, )
 				console.log('name field was updated', updatedChannel.name === 'updated')
 				const {data: track} = await sdk.track.createTrack(channel.id, {url: 'https://www.youtube.com/watch?v=dA55o_18a-g', title: 'My new track'})
