+++ src/index.js
@@ -1,7 +1,7 @@
 import * as auth from './auth.js'
-import * as user from './user.js'
-import * as channel from './channel.js'
-import * as track from './track.js'
+import * as user from './users.js'
+import * as channel from './channels.js'
+import * as track from './tracks.js'
 
 export {mediaUrlParser} from 'media-url-parser'
 export {supabase} from './supabase-client.js'
@@ -28,6 +28,6 @@ export {
 }
 
 export * from './auth.js'
-export * from './user.js'
-export * from './channel.js'
-export * from './track.js'
+export * from './users.js'
+export * from './channels.js'
+export * from './tracks.js'
